# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Growthly Platform API Docs",
    "description": "",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Local API Development"
    }
  ],
  "paths": {
...
    "/auth/update-2fa": {
      "post": {
        "summary": "Enable or disable Two-Factor Authentication (2FA)",
        "tags": ["Authentication"],
        "description": "Updates the 2FA status for the authenticated user based on the `action` provided in the request body. Requires a full session (email must be verified).",
        "operationId": "update2FA",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "action": {
                    "type": "string",
                    "enum": ["enable", "disable"],
                    "example": "enable",
                    "description": "Set to `enable` to turn on 2FA or `disable` to turn it off"
                  }
                },
                "required": ["action"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "2FA status updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Two-factor authentication has been enabled."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid action provided or 2FA already in requested state",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid action. Use \"enable\" or \"disable\"."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized, No token provided"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - partial sessions cannot modify 2FA",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You are not allowed to perform this action"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get basic profile of a user by ID",
        "tags": [
          "Users"
        ],
        "description": "Requires the user to be authenticated. Returns public profile information of a user, including ID, name, email, role, and associated links. Sensitive fields such as passwords and authentication tokens are excluded.",
        "operationId": "getUserById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "The unique identifier of the user.",
            "schema": {
              "type": "string",
              "example": "johndoe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's profile information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "johndoe@gmail.com"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "headline": {
                          "type": "string",
                          "example": "Software Engineer"
                        },
                        "bio": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": null
                        },
                        "country": {
                          "type": "string",
                          "example": "USA"
                        },
                        "imageUrl": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": null
                        },
                        "dateJoined": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-07-11T16:58:26.127Z"
                        },
                        "skills": {
                          "type": [
                            "array",
                            "null"
                          ],
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "HTML",
                            "CSS",
                            "JavaScript"
                          ]
                        },
                        "timezone": {
                          "type": "string",
                          "example": "Etc/UTC"
                        },
                        "role": {
                          "type": "string",
                          "example": "MENTOR"
                        },
                        "links": {
                          "type": [
                            "array",
                            "null"
                          ],
                          "items": {
                            "type": "object",
                            "properties": {
                              "name": {
                                "type": "string",
                                "example": "GitHub"
                              },
                              "url": {
                                "type": "string",
                                "format": "uri",
                                "example": "https://github.com/johndoe"
                              }
                            }
                          },
                          "example": [
                            {
                              "name": "GitHub",
                              "url": "https://github.com/johndoe"
                            },
                            {
                              "name": "Portfolio",
                              "url": "https://johndoe.dev"
                            }
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized, No token provided"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "User not found"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get current authenticated user's profile",
        "tags": [
          "Users"
        ],
        "description": "Requires the user to be authenticated. Returns basic user information including ID, name, email, role, and other optional profile fields.",
        "operationId": "getCurrentUser",
        "responses": {
          "200": {
            "description": "Successfully retrieved the user's profile information.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "johndoe@gmail.com"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "headline": {
                          "type": "string",
                          "example": "Software Engineer"
                        },
                        "bio": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": null
                        },
                        "country": {
                          "type": "string",
                          "example": "USA"
                        },
                        "imageUrl": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": null
                        },
                        "dateJoined": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-07-11T16:58:26.127Z"
                        },
                        "skills": {
                          "type": [
                            "array",
                            "null"
                          ],
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "HTML",
                            "CSS",
                            "JavaScript"
                          ]
                        },
                        "timezone": {
                          "type": "string",
                          "example": "Etc/UTC"
                        },
                        "role": {
                          "type": "string",
                          "example": "MENTOR"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized, No token provided"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update authenticated user's profile",
        "tags": [
          "Users"
        ],
        "description": "Requires the user to be authenticated. Allows the user to update their profile information. At least one updatable field must be provided.",
        "operationId": "updateAuthenticatedUser",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 50,
                    "example": "John Doe"
                  },
                  "headline": {
                    "type": "string",
                    "maxLength": 50,
                    "example": "Full-Stack Developer"
                  },
                  "bio": {
                    "type": "string",
                    "maxLength": 1000,
                    "example": "I'm passionate about clean code and teaching developers."
                  },
                  "country": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 3,
                    "description": "A valid ISO 3166-1 alpha-3 country code",
                    "example": "USA"
                  },
                  "timezone": {
                    "type": "string",
                    "description": "A valid IANA time zone name",
                    "example": "America/New_York"
                  },
                  "skills": {
                    "type": "array",
                    "maxItems": 300,
                    "items": {
                      "type": "string",
                      "maxLength": 50
                    },
                    "example": [
                      "JavaScript",
                      "React",
                      "Node.js"
                    ]
                  }
                },
                "example": {
                  "headline": "Backend Engineer",
                  "bio": "10+ years in distributed systems",
                  "skills": [
                    "Node.js",
                    "Docker",
                    "PostgreSQL"
                  ]
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the user's profile.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Success"
                    },
                    "message": {
                      "type": "string",
                      "example": "User updated successfully"
                    },
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "example": "johndoe"
                        },
                        "email": {
                          "type": "string",
                          "format": "email",
                          "example": "johndoe@gmail.com"
                        },
                        "name": {
                          "type": "string",
                          "example": "John Doe"
                        },
                        "headline": {
                          "type": "string",
                          "example": "Backend Engineer"
                        },
                        "bio": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": "10+ years in distributed systems"
                        },
                        "country": {
                          "type": "string",
                          "example": "USA"
                        },
                        "imageUrl": {
                          "type": [
                            "string",
                            "null"
                          ],
                          "example": null
                        },
                        "dateJoined": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2025-07-11T16:58:26.127Z"
                        },
                        "skills": {
                          "type": [
                            "array",
                            "null"
                          ],
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "Node.js",
                            "Docker",
                            "PostgreSQL"
                          ]
                        },
                        "timezone": {
                          "type": "string",
                          "example": "America/New_York"
                        },
                        "role": {
                          "type": "string",
                          "example": "MENTOR"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error or no valid fields provided",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "At least 1 attribute should be provided to update"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - user is not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "Fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Unauthorized, No token provided"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}